Get base upstream commit from gitrev
We added a file in 4.3.0-0.rc3.git2.1.fc24 called 'gitrev' that lists the
exact upstream commit sha1sum that the git snapshot was created from.  We
read this file to get the commit sha1sum to use as the base of our exploded
tree.

This means the scripts won't handle git snapshot kernels before the listed
version, but there isn't much we can do about that without a lot of work.