Fix fedpkg client usage
Apparently the fedpkg/pyrpkg code caches values it determines when running
the first command.  That means that we'd get a successful run on the first
'fedpkg prep' we do from the code, but then the values derived from the
pkg-git repo at that time are saved.  When the next build comes in, the
pkg-git repo branch would be switched but the prep command run on it would
use the disttag value from the previous run.  That led to all kinds of
path problems a few lines later.

Fix this by just creating the client instance in the prep_pkg_git function
every time.  There's really no other use of the fedpkg client code anywhere
else, so this cleans things up a bit as well.