Parse the major kernel version
We need to provide the major kernel version from the spec as well.
This only changes once in a very long while, but we might as well
account for it.

On a major version changeover, the kernel team uses RC tarballs
instead of prior release tarballs and rc patches.  This throws
the tooling for a bit of a loop, so also look to see if there's
a tarball "suffix" used when it's exploded.  E.g.

kernel-5.0-rc1/