Update to use major release values from the spec
Make sure we account for the unique situation of using RC tarballs
instead of prior release tarballs plus RC patches.  This really
only happens on new major .0 releases, but it's good to try and be
consistent.

If the kernel team changes approach in the future, we'll need to
update this code.  It has to rely on some assumptions on how the
kernel team operates.